feat(gh): merge workflows and run only for PRs
We have been running the same workflows twice for a PR as there has been
a push trigger without any branch condition. As this is unecessary and
the structure a bit convoluted anyway, this integrates everything into a
single workflow for PRs to build and test in strict order and only for
pull requests. Publishing is a different step and will be handled by the
release workflow.