fix(js-sdk): cancel video frame callback when video stream stops
When video upscaling is enabled, the video frame is rendered in the
way based on the HTMLVideoElement.requestVideoFrameCallback function.
However, due to a missing call to HTMLVideoElement.cancelVideoFrameCallback,
the AnboxStreamCanvas._render function continued to be called for
small period of time even after the video stream had stopped.

To resolve this issue, ensure that any registered video frame callbacks
are canceled when the video stream stops.