feat(js-sdk): use `#version 300 es` for all built-in shaders
In WebGL, the vertex shader and fragment shader must use the same
GLSL version, otherwise it will cause a version mismatch error
```
Fragment shader version does not match other shader versions.
```

for now, we use `#version 100` for the built-in vertex shader, however
if the provided the fragment shdaer is using extension or features
from `#version 300 es`, it won't work with the JS SDK. Hence
for the forward compatibility, we use `#version 300 es` for all
built-in shaders.

Also because we bumped up the GLES version to `#version 300 es`,
the following deprecated GL directives can be dropped now. E.g.
- gl_FragColor
- attribute
- varying
- texture2D

Test: enable the upscaling feature and video streaming works fine