feat(js-sdk): add and integrate Snapdragon Game Super Resolution(SGSR) into js exmaples
Snapdragon™ Game Super Resolution(SGSR), which integrates upscaling
and sharpening in one single GPU shader pass. The algorithm uses a
12-tap Lanczos-like scaling filter and adaptive sharpening filter,
which presents smooth images and sharp edges. For more details:

 https://github.com/SnapdragonStudios/snapdragon-gsr

This change adds and integrate SGSR, this single GPU shader pass
into the streaming js sdk, and usd by default for the time being.

NOTE: FSR is still kept in the source for reference. In the end,
we will choose the best upscaling algorithm from all available
options after benchmarking and comparing performance and quality,
and only ship the best one in the streaming SDK. Hence we ship this
as an example as of now.