feat(js-sdk): provide an additional uniform for the source video texture
For some multi-pass shaders used for upscaling (e.g., Anime4K),
various operations are performed on the source image, such as edge
detection and color adjustment. And the processed image is blended
with the original image to generate a clearer result. For example:

Anime4K Upscale Denoise Shader
https://github.com/bloc97/Anime4K/blob/master/glsl/Upscale%2BDenoise/Anime4K_Upscale_Denoise_CNN_x2_S.glsl#L143

This change adds support for a new uniform that denotes the source
video texture, which can be used by a shader to blend the processed
image with the original video source.