fix(e2e): do not test AP_POWER_STATE_REPORT VHAL prop temporarily
Due to a permission issue, the e2e test that changes the
AP_POWER_STATE_REPORT VHAL prop is broken. The issue has been fixed
upstream, but since it will take a few days before the new image is
published, we temporarily disable the test for this VHAL prop, and
introduce a new one that acts on the same value type (array of
integers), for the WHEEL_TICK prop, that does not have permission
issues, so that the CI is not blocked for a long time.