fix(examples): remove the faulty note from the README of the out_of_band_v2 example
The note in the README file of the out_of_band_v2 example is
incorrect. To install the resulting out_of_band_v2 APK as a
system app, the APK must be shipped into the Android root filesystem
using a pre-start hook as documented in the following page,

  https://documentation.ubuntu.com/anbox-cloud/en/latest/howto/port/install-apk-system-app/

since the custom system signature cannot be injected during the
Android runtime but must be done prior to Android startup.