fix(js): do not abort streaming on non-fatal errors
If a client accidentally enables the camera while their device is
not equipped with one, it triggers the `ANBOX_STREAM_SDK_ERROR_USER_MEDIA`
error. However, this is not a fatal error and make less sense to
abort the streaming when no camera is found.

In such cases, we simply report the error to the end user and
prevent the streaming SDK from handling it.