fix(js): do not send data to the server if there is no data change
Currently, the _onSensorDataUpdate() is based on time throttling,
but it does not check whether the sensor data has really changed.
As a result, even if the data remains exactly the same, it will
still be sent.

The sensor manager on the server side has a worker thread that consumes
the latest cached data, so sending identical data is unnecessary
it wastes bandwidth and adds extra parsing load on the server.